This brach contains refactoring of the middlewares and file-generator that fix following issue:
Let's assume that we have a node that we decided to extract to a component and this node should be center aligned regardless its parent.
The problem was that in layout-and-size middleware we wrap this node inside a View. View will take all available space and node will position at center of the View.
All this logic will be stored inside `renderCode` function.

Next, inside export-to-component middleware we will decide that component should be extracted to a component, so we will create a file and render code using `renderCode` function,
but this function also contains render of View and center positioned element, that should not be part of the new component. We would like to extract to a component only node,
and wrap by View only instance of that component, but we can't because we have only one function 'renderCode'.

To avoid this I decided to separate node rendering into two parts:
`importComponent` and `renderComponent` will take care to rendering component only (everything that should be extracted inside a component)
`importDecorator` and `renderDecorator` will take care to rendering everything that wrap the component (layout or 'wrapWith' extend)